test the dao with update and delete

+++ src/main/java/lk/ijse/aad/gdse68/notetaker/service/NoteServiceImpl.java
@@ -9,11 +9,8 @@ import lk.ijse.aad.gdse68.notetaker.util.Mapping;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
 import java.util.List;
-import java.util.ListIterator;
-
-import static org.antlr.v4.runtime.tree.xpath.XPath.findAll;
+import java.util.Optional;
 
 @Service
 @Transactional
@@ -32,12 +29,26 @@ public class NoteServiceImpl implements NoteService {
 
     @Override
     public boolean updateNote(String noteId, NoteDto incomenoteDto) {
-        return false;
+        Optional<NoteEntity> updateById=noteDao.findById(noteId);
+        if(!updateById.isPresent()){
+            return false;
+        }else {
+            updateById.get().setNoteDecs(incomenoteDto.getNoteDecs());
+            updateById.get().setNoteTitle(incomenoteDto.getNoteTitle());
+            updateById.get().setCreateDate(incomenoteDto.getCreateDate());
+            updateById.get().setPriorityLevel(incomenoteDto.getPriorityLevel());
+        return true;
+        }
     }
 
     @Override
-    public void deleteNote(String noteId) {
-    noteDao.deleteById(noteId);
+    public boolean deleteNote(String noteId) {
+        if(noteDao.existsById(noteId)){
+            noteDao.deleteById(noteId);
+            return true;
+        }
+        return false;
+
     }
     @Override
     public NoteDto getSelectNote(String noteId) {
